client/backend/requirements.txt
Flask==2.0.2
Flask_Cors==3.0.10
mysql_connector_python==8.0.27
